Fix nightly to the 'demos are green' condition

+++ src/ioreg/test.rs
@@ -15,9 +15,10 @@
 
 //! Tests for ioreg! syntax extension
 
-#![feature(phase)]
-#[phase(plugin)] extern crate ioreg;
-#[phase(plugin,link)] extern crate shiny;
+#![feature(plugin)]
+
+#[plugin] extern crate ioreg;
+#[plugin] extern crate shiny;
 extern crate core;
 
 #[path="../zinc/util/volatile_cell.rs"] mod volatile_cell;
@@ -28,10 +29,10 @@ mod test {
   use std::ptr::PtrExt;
   use volatile_cell::VolatileCell;
 
-  fn get_value<'a, T>(v: &'a T, offset: uint) -> u32 {
+  fn get_value<'a, T>(v: &'a T, offset: usize) -> u32 {
     unsafe {
       let ptr: *const u32 = transmute(v);
-      *(ptr.offset(offset as int))
+      *(ptr.offset(offset as isize))
     }
   }
 
@@ -162,15 +163,15 @@ mod test {
     }
     it "has zero base offset" {
       let addr = &test.reg1 as *const GAP_TEST_reg1;
-      assert_eq!(addr as uint - base as uint, 0x0);
+      assert_eq!(addr as usize - base as usize, 0x0);
     }
     it "computes the correct first gap" {
       let addr = &test.reg2 as *const GAP_TEST_reg2;
-      assert_eq!(addr as uint - base as uint, 0x10);
+      assert_eq!(addr as usize - base as usize, 0x10);
     }
     it "computes the correct second gap" {
       let addr = &test.reg4 as *const GAP_TEST_reg4;
-      assert_eq!(addr as uint - base as uint, 0x20);
+      assert_eq!(addr as usize - base as usize, 0x20);
     }
   );
 }
